pip_requirements.txt
absl-py==0.9.0
asn1crypto==0.24.0
backcall==0.1.0
beautifulsoup4==4.8.0
cachetools==4.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
conda==4.7.11
conda-build==3.18.9
conda-package-handling==1.3.11
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
filelock==3.0.12
future==0.18.2
glob2==0.7
google-auth==1.17.2
google-auth-oauthlib==0.4.1
grpcio==1.29.0
idna==2.8
imageio==2.8.0
importlib-metadata==1.6.1
interpret-segmentation==1.0.0
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
Jinja2==2.10.1
joblib==0.15.1
kiwisolver==1.2.0
libarchive-c==2.8
lief==0.9.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mkl-fft==1.0.14
mkl-random==1.0.2
mkl-service==2.3.0
networkx==2.4
numpy==1.16.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.34
pandas==1.0.4
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pkginfo==1.5.0.1
prompt-toolkit==2.0.9
protobuf==3.12.2
psutil==5.6.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser==2.19
Pygments==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.7
PySocks==1.7.0
python-dateutil==2.8.1
pytz==2019.2
PyWavelets==1.1.1
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel-yaml==0.15.46
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
six==1.12.0
sklearn==0.0
soupsieve==1.9.2
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.0
threadpoolctl==2.1.0
tifffile==2020.6.3
torch==1.2.0
torchvision==0.4.0a0+6b959ee
tqdm==4.32.1
traitlets==4.3.2
urllib3==1.24.2
wcwidth==0.1.7
Werkzeug==1.0.1
zipp==3.1.0
